Modified both the LightCurveDataset and FeetsFeatureExctractor classes. Tested Astronomaly on PlAsTiCC data (100 objects).
