+++ astronomaly/feature_extraction/feets_features.py
@@ -39,7 +39,7 @@ class Feets_Features(PipelineStage):
             incase there is an error during the feature extraction process'''
 
         # Sorting the columns for the feature extractor
-
+        print(lc_data.ID)
         # This needs to be extended to be more general
         standard_lc_columns = ['time', 'mag', 'mag_error']
         current_lc_columns = [cl for cl in standard_lc_columns
@@ -72,15 +72,36 @@ class Feets_Features(PipelineStage):
             lc_columns = []
             for col in current_lc_columns:
                 lc_columns.append(lc_data[col])
-            features, values = fs.extract(*lc_columns)
 
-            # Updating the labels
+            if 'filters' in lc_data.columns:
+                ft_values = []
+                ft_labels = []
+                for i in range(len(np.unique(lc_data['filters']))):
+
+                    passbands = ['u', 'g', 'r', 'i', 'z', 'y']
+                    lc_data = lc_data[lc_data['filters'] == i]
+                    features, values = fs.extract(*lc_columns)
+
+                    new_labels = [f + '_' + passbands[i] for f in features]
+
+                    for j in range(len(features)):
+                        ft_labels.append(new_labels[j])
+                        ft_values.append(values[j])
+
+                if np.nan in ft_values:
+                    nan = [np.nan for i in range(len(ft_values))]
+                    ft_values = np.array(nan)
+
+            else:
+                ft_labels, ft_values = fs.extract(*lc_columns)
+
+            # # Updating the labels
             if self.labels is None:
 
-                self._set_labels(list(features))
+                self._set_labels(list(ft_labels))
 
             # The calculated features
-            return values
+            return ft_values
 
         # Returns an array of nan values
         else:
